try middle struct
